+++ src/core/loader_tmx.cpp
@@ -116,7 +116,7 @@ void tmx_class::load(std::string file_name)
                             temp_string_key += temp_char;
                         }
                         position_count++;
-                        if(position_count > data_line.length()) temp_char = '>';
+                        if(position_count > (int)data_line.length()) temp_char = '>';
                     }
                     if (temp_string_key == "tileset")
                     {
@@ -127,7 +127,7 @@ void tmx_class::load(std::string file_name)
                 if((data_line[position_count] == '<') && (data_line[position_count+1] != '/') && (data_line.length() > 2))
                 {
                     position_count++;
-                    while(position_count < data_line.length())
+                    while(position_count < (int)data_line.length())
                     {
                         temp_char         = ' ';
                         temp_string_key   = "";
@@ -141,7 +141,7 @@ void tmx_class::load(std::string file_name)
                                 else if (temp_char != ' ') temp_string_key += temp_char;
                             }
                             position_count++;
-                            if(position_count > data_line.length()) temp_char = '"';
+                            if(position_count > (int)data_line.length()) temp_char = '"';
                         }
                         temp_char        = ' ';
                         while(temp_char != '"')
@@ -149,7 +149,7 @@ void tmx_class::load(std::string file_name)
                             temp_char = data_line[position_count];
                             if(temp_char != '"') temp_string_value += temp_char;
                             position_count++;
-                            if(position_count > data_line.length()) (temp_char = '"');
+                            if(position_count > (int)data_line.length()) (temp_char = '"');
                         }
                         temp_string_data    = temp_string_value.c_str();
                         temp_float_data     = atof(temp_string_value.c_str());
@@ -217,7 +217,7 @@ void tmx_class::load(std::string file_name)
                             temp_string_key += temp_char;
                         }
                         position_count++;
-                        if(position_count > data_line.length()) temp_char = '>';
+                        if(position_count > (int)data_line.length()) temp_char = '>';
                     }
                     if (temp_string_key == "tileset")
                     {
@@ -239,7 +239,7 @@ void tmx_class::load(std::string file_name)
                 if((data_line[position_count] == '<') && (data_line[position_count+1] != '/') && (data_line.length() > 2))
                 {
                     position_count++;
-                    while(position_count < data_line.length())
+                    while(position_count < (int)data_line.length())
                     {
                         temp_char         = ' ';
                         temp_string_key   = "";
@@ -253,7 +253,7 @@ void tmx_class::load(std::string file_name)
                                 else if (temp_char != ' ') temp_string_key += temp_char;
                             }
                             position_count++;
-                            if(position_count > data_line.length()) temp_char = '"';
+                            if(position_count > (int)data_line.length()) temp_char = '"';
                         }
                         temp_char        = ' ';
                         while(temp_char != '"')
@@ -261,7 +261,7 @@ void tmx_class::load(std::string file_name)
                             temp_char = data_line[position_count];
                             if(temp_char != '"') temp_string_value += temp_char;
                             position_count++;
-                            if(position_count > data_line.length()) (temp_char = '"');
+                            if(position_count > (int)data_line.length()) (temp_char = '"');
                         }
                         temp_string_data    = temp_string_value.c_str();
                         temp_float_data     = atof(temp_string_value.c_str());
